Move netlink files to network directory
