+++ src/data/os/linux.cc
@@ -87,7 +87,7 @@
 #endif
 
 #ifdef BUILD_WLAN
-#include "netlink-conky.h"
+#include "../network/netlink-conky.h"
 #endif
 
 struct sysfs {
